requirements.txt
# Core Golem DB dependencies
golem-base-sdk>=0.1.0
web3>=7.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.0.0

# GUI dependencies
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Data analysis and visualization
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0

# Blockchain data fetching
websockets>=11.0.0


